feat(promotion): 할인 정책 추가
