--- src/main/java/config/DependencyInjector.java
@@ -1,10 +1,0 @@
-package config;
-
-import controller.Controller;
-
-public class DependencyInjector {
-    public Controller createController() {
-
-        return new Controller();
-    }
-}
