--- src/main/java/controller/Controller.java
@@ -1,8 +1,0 @@
-package controller;
-
-public class Controller {
-
-    public void run() {
-    }
-
-}
