--- src/main/java/controller/InputParser.java
@@ -1,13 +1,0 @@
-package controller;
-
-public class InputParser {
-
-    public static int parseInt(String input) {
-        try {
-            return Integer.parseInt(input);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(e);
-        }
-    }
-
-}
