+++ src/main/java/cristmas/config/DependencyInjector.java
@@ -1,0 +1,10 @@
+package cristmas.config;
+
+import cristmas.controller.Controller;
+
+public class DependencyInjector {
+    public Controller createController() {
+
+        return new Controller();
+    }
+}
