+++ src/main/java/cristmas/controller/Controller.java
@@ -1,0 +1,8 @@
+package cristmas.controller;
+
+public class Controller {
+
+    public void run() {
+    }
+
+}
