+++ src/main/java/cristmas/controller/InputParser.java
@@ -1,0 +1,13 @@
+package cristmas.controller;
+
+public class InputParser {
+
+    public static int parseInt(String input) {
+        try {
+            return Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException(e);
+        }
+    }
+
+}
