+++ src/main/java/cristmas/domain/product/ProductField.java
@@ -1,0 +1,7 @@
+package cristmas.domain.product;
+
+
+import cristmas.enums.ProductType;
+
+public record ProductField(String name, int price, ProductType type) {
+}
