+++ src/main/java/cristmas/enums/DayOfWeek.java
@@ -1,0 +1,11 @@
+package cristmas.enums;
+
+public enum DayOfWeek {
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY,
+    SUNDAY
+}
