+++ src/main/java/cristmas/enums/ProductType.java
@@ -1,0 +1,8 @@
+package cristmas.enums;
+
+public enum ProductType {
+    APPETIZER,
+    MAIN,
+    DESERT,
+    DRINK,
+}
