+++ src/main/java/cristmas/exception/GlobalErrorMessage.java
@@ -1,0 +1,15 @@
+package cristmas.exception;
+
+public enum GlobalErrorMessage {
+    ;
+
+    private final String message;
+
+    GlobalErrorMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
