+++ src/main/java/cristmas/file/parser/GenericParser.java
@@ -1,0 +1,7 @@
+package cristmas.file.parser;
+
+import java.io.IOException;
+
+public interface GenericParser<T> {
+    T nextLine() throws IOException;
+}
