+++ src/main/java/cristmas/file/parser/ProductParser.java
@@ -1,0 +1,12 @@
+package cristmas.file.parser;
+
+import cristmas.domain.product.ProductField;
+import cristmas.file.reader.Reader;
+
+public class ProductParser extends AbstractFieldParser<ProductField> {
+
+    public ProductParser(Reader reader) {
+        super(reader, ProductField.class);
+    }
+
+}
