+++ src/main/java/cristmas/file/reader/Reader.java
@@ -1,0 +1,10 @@
+package cristmas.file.reader;
+
+import java.io.IOException;
+import java.util.List;
+
+public interface Reader {
+    List<String> readLine() throws IOException;
+
+    void close() throws IOException;
+}
