+++ src/main/java/cristmas/view/InputView.java
@@ -1,0 +1,9 @@
+package cristmas.view;
+
+public class InputView {
+
+//    public static String getUserInput() {
+//        return Console.readline();
+//    }
+
+}
