+++ src/main/java/cristmas/view/OutputView.java
@@ -1,0 +1,9 @@
+package cristmas.view;
+
+public class OutputView {
+
+    public static void printError(String message) {
+        System.out.println("[ERROR] " + message);
+        System.out.println();
+    }
+}
