--- src/main/java/domain/product/ProductField.java
@@ -1,5 +1,0 @@
-package domain.product;
-
-
-public record ProductField(String name, int price, ProductType type) {
-}
