--- src/main/java/domain/product/ProductType.java
@@ -1,8 +1,0 @@
-package domain.product;
-
-public enum ProductType {
-    APPETIZER,
-    MAIN,
-    DESERT,
-    DRINK,
-}
