--- src/main/java/exception/GlobalErrorMessage.java
@@ -1,15 +1,0 @@
-package exception;
-
-public enum GlobalErrorMessage {
-    ;
-
-    private final String message;
-
-    GlobalErrorMessage(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
