--- src/main/java/file/parser/GenericParser.java
@@ -1,7 +1,0 @@
-package file.parser;
-
-import java.io.IOException;
-
-public interface GenericParser<T> {
-    T nextLine() throws IOException;
-}
