--- src/main/java/file/reader/Reader.java
@@ -1,10 +1,0 @@
-package file.reader;
-
-import java.io.IOException;
-import java.util.List;
-
-public interface Reader {
-    List<String> readLine() throws IOException;
-
-    void close() throws IOException;
-}
