--- src/main/java/view/InputView.java
@@ -1,9 +1,0 @@
-package view;
-
-public class InputView {
-
-//    public static String getUserInput() {
-//        return Console.readline();
-//    }
-
-}
