--- src/main/java/view/OutputView.java
@@ -1,9 +1,0 @@
-package view;
-
-public class OutputView {
-
-    public static void printError(String message) {
-        System.out.println("[ERROR] " + message);
-        System.out.println();
-    }
-}
